Changed Password storage to plain text storage
This change will make the app a slight bit more secure (no more hardcoded passwords) but still is not secure as the password is stored in plain text ("pass.txt").

Additionally, the code was cleaned up. Firstly, a new prompt_pass! macro was created that had the affect of drastically drying up code. Secondly, password authentication was removed from list(), new(), and remove(). This responsibility was delegated to a new function, authenticate(), which will handle the password logic.

Signed-off-by: BenRaz123 <ben.raz2008@gmail.com>