Reached Feature Parity with master
The code uses the sqlite library.

The architecture is as follows: two tables, one for the password, the other for the secrets. The password table is as follows:

| PASSWORDS  |
|------------|
| {PASSWORD} |

The password database is cleared whenever a new password is inserted, keeping a constant length of 1.

The the secrets table is as follows:

| ID  | SECRETS |
|-----|---------|
| u64 | String  |

Signed-off-by: BenRaz123 <ben.raz2008@gmail.com>